Seperate prediction and camera logic into seperate hooks
