+++ client/src/components/FridgeItem/FridgeItem.jsx
@@ -3,11 +3,10 @@ import './FridgeItem.scss';
 import { BsCart4, BsTrash } from 'react-icons/bs';
 import axios from 'axios';
 import swal from 'sweetalert';
-import { useNavigate, useParams } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 import classNames from "classnames";
 
 const FridgeItem = (props) => {
-  const { id } = useParams();
   const navigate = useNavigate();
 
   const daysAgo = (days) => {
