+++ client/src/hooks/useCamera/useCamera.js
@@ -1,4 +1,4 @@
-import { useState, useEffect } from "react";
+import { useState } from 'react';
 
 const defaultConstraints = {
   audio: false,
@@ -7,22 +7,38 @@ const defaultConstraints = {
   height: { ideal: 1080 }
 }
 
-export default function useCamera(video, onLoad, constraints = defaultConstraints) {
+export default function useCamera(videoRef, canvasRef, constraints = defaultConstraints) {
+  const [picture, setPicture] = useState();
 
-
-
-  useEffect(() => {
+  const requestUserMedia = (onLoad) => {
     // Check if client supports navigator mediaDevices and getUserMedia API
     if (navigator.mediaDevices && navigator.mediaDevices.getUserMedia) {
       navigator.mediaDevices
         .getUserMedia(constraints)
         .then(stream => {
-          video.current.srcObject = stream;
-          console.log(video.current);
-          video.current.onloadedmetadata = function() {
+          videoRef.current.srcObject = stream;
+          videoRef.current.onloadedmetadata = function() {
             onLoad();
           };
         })
     }
-  }, [video, onLoad, constraints]);
+  };
+
+  const takePicture = (cb) => {
+    const canvas = canvasRef.current;
+    canvas.width = 640;
+    canvas.height = 480;
+    const context = canvas.getContext('2d');
+    const video = document.getElementsByClassName("camera__video")[0];
+    context.drawImage(video, 0, 0, canvas.width, canvas.height);
+    canvas.toBlob(img => {
+      setPicture(img);
+      context.clearRect(0, 0, canvas.width, canvas.height);
+      if(cb) {
+        cb(img);
+      }
+    })
+  };
+
+  return { picture, takePicture, requestUserMedia }
 }
